Debug Game.
The PR to merge contained issues where some code still needed to be refactored in PapaChad.js to use the vector class functionality. The game was loading Village which is not in the same adequate playable state as PlayGround. I updated the code in PapaChad to use "pos" where necessaary with the new Vector class. The dash functionality was broken so I removed code modularity due to a scope issue with xVelocity variable and put the functionality of DashAction back into the update loop code and removed a buggy portion of DashAction. I commented out some console.log code in slinshot that used undefined variables.